+++ tests/testing/test_model_builder.py
@@ -51,6 +51,10 @@ class BandWithLazyReference(Table):
     )
 
 
+class BandWithRecursiveReference(Table):
+    manager: ForeignKey["Manager"] = ForeignKey("self")
+
+
 TABLES = (
     Manager,
     Band,
@@ -63,6 +67,7 @@ TABLES = (
     TableWithArrayField,
     TableWithDecimal,
     BandWithLazyReference,
+    BandWithRecursiveReference,
 )
 
 
@@ -93,6 +98,7 @@ class TestModelBuilder(unittest.TestCase):
         queried_shirt = (
             Shirt.objects().where(Shirt.id == shirt.id).first().run_sync()
         )
+        assert queried_shirt is not None
 
         self.assertIn(
             queried_shirt.size,
@@ -132,6 +138,16 @@ class TestModelBuilder(unittest.TestCase):
             Manager.exists().where(Manager.id == model.manager).run_sync()
         )
 
+    def test_recursive_foreign_key(self):
+        """
+        Make sure no infinite loops are created with recursive foreign keys.
+        """
+        model = ModelBuilder.build_sync(
+            BandWithRecursiveReference, persist=True
+        )
+        # It should be set to None, as this foreign key is nullable.
+        self.assertIsNone(model.manager)
+
     def test_invalid_column(self):
         with self.assertRaises(ValueError):
             ModelBuilder.build_sync(Band, defaults={"X": 1})
@@ -157,6 +173,7 @@ class TestModelBuilder(unittest.TestCase):
             .first()
             .run_sync()
         )
+        assert queried_manager is not None
 
         self.assertEqual(queried_manager.name, "Guido")
 
@@ -169,6 +186,7 @@ class TestModelBuilder(unittest.TestCase):
             .first()
             .run_sync()
         )
+        assert queried_manager is not None
 
         self.assertEqual(queried_manager.name, "Guido")